+++ src/anomalib/data/image/visa.py
@@ -51,7 +51,7 @@ EXTENSIONS = (".png", ".jpg", ".JPG")
 DOWNLOAD_INFO = DownloadInfo(
     name="VisA",
     url="https://amazon-visual-anomaly.s3.us-west-2.amazonaws.com/VisA_20220922.tar",
-    checksum="ef908989b6dc701fc218f643c127a4de",
+    hashsum="2eb8690c803ab37de0324772964100169ec8ba1fa3f7e94291c9ca673f40f362",
 )
 
 CATEGORIES = (
@@ -196,12 +196,12 @@ class Visa(AnomalibDataModule):
         train_batch_size: int = 32,
         eval_batch_size: int = 32,
         num_workers: int = 8,
-        task: TaskType = TaskType.SEGMENTATION,
+        task: TaskType | str = TaskType.SEGMENTATION,
         transform_config_train: str | A.Compose | None = None,
         transform_config_eval: str | A.Compose | None = None,
-        test_split_mode: TestSplitMode = TestSplitMode.FROM_DIR,
+        test_split_mode: TestSplitMode | str = TestSplitMode.FROM_DIR,
         test_split_ratio: float = 0.2,
-        val_split_mode: ValSplitMode = ValSplitMode.SAME_AS_TEST,
+        val_split_mode: ValSplitMode | str = ValSplitMode.SAME_AS_TEST,
         val_split_ratio: float = 0.5,
         seed: int | None = None,
     ) -> None:
@@ -234,14 +234,14 @@ class Visa(AnomalibDataModule):
         )
 
         self.train_data = VisaDataset(
-            task=task,
+            task=TaskType(task),
             transform=transform_train,
             split=Split.TRAIN,
             root=self.split_root,
             category=category,
         )
         self.test_data = VisaDataset(
-            task=task,
+            task=TaskType(task),
             transform=transform_eval,
             split=Split.TEST,
             root=self.split_root,