+++ src/anomalib/engine/engine.py
@@ -12,7 +12,6 @@ from lightning.pytorch.callbacks import Callback
 from lightning.pytorch.trainer import Trainer
 from lightning.pytorch.trainer.connectors.callback_connector import _CallbackConnector
 from lightning.pytorch.utilities.types import _EVALUATE_OUTPUT, _PREDICT_OUTPUT, EVAL_DATALOADERS, TRAIN_DATALOADERS
-from omegaconf import DictConfig, ListConfig
 from torch.utils.data import DataLoader, Dataset
 
 from anomalib import LearningType, TaskType
@@ -25,9 +24,9 @@ from anomalib.callbacks.visualizer import _VisualizationCallback
 from anomalib.data import AnomalibDataModule, AnomalibDataset, PredictDataset
 from anomalib.data.utils import TestSplitMode
 from anomalib.deploy.export import ExportType, export_to_onnx, export_to_openvino, export_to_torch
-from anomalib.metrics.threshold import BaseThreshold
 from anomalib.models import AnomalyModule
 from anomalib.utils.normalization import NormalizationMethod
+from anomalib.utils.types import NORMALIZATION, THRESHOLD
 from anomalib.utils.visualization import BaseVisualizer
 
 logger = logging.getLogger(__name__)
@@ -35,8 +34,6 @@ logger = logging.getLogger(__name__)
 
 class UnassignedError(Exception):
     """Unassigned error."""
-
-    ...
 
 
 class _TrainerArgumentsCache:
@@ -106,16 +103,16 @@ class Engine:
 
     Args:
         callbacks (list[Callback]): Add a callback or list of callbacks.
-        normalization (NormalizationMethod | DictConfig | Callback | str, optional): Normalization method.
+        normalization (NORMALIZATION, optional): Normalization method.
             Defaults to NormalizationMethod.MIN_MAX.
-        threshold (BaseThreshold | tuple[BaseThreshold, BaseThreshold] | DictConfig | ListConfig | str, optional):
+        threshold (THRESHOLD):
             Thresholding method. Defaults to "F1AdaptiveThreshold".
         task (TaskType, optional): Task type. Defaults to TaskType.SEGMENTATION.
         image_metrics (str | list[str] | None, optional): Image metrics to be used for evaluation.
             Defaults to None.
         pixel_metrics (str | list[str] | None, optional): Pixel metrics to be used for evaluation.
             Defaults to None.
-        visualization_handlers (BaseVisualizationGenerator | list[BaseVisualizationGenerator] | None):
+        visualizers (BaseVisualizationGenerator | list[BaseVisualizationGenerator] | None):
             Visualization parameters. Defaults to None.
         **kwargs: PyTorch Lightning Trainer arguments.
     """
@@ -123,14 +120,9 @@ class Engine:
     def __init__(
         self,
         callbacks: list[Callback] | None = None,
-        normalization: NormalizationMethod | DictConfig | Callback | str = NormalizationMethod.MIN_MAX,
-        threshold: BaseThreshold
-        | tuple[BaseThreshold, BaseThreshold]
-        | DictConfig
-        | ListConfig
-        | list[dict[str, str | float]]
-        | str = "F1AdaptiveThreshold",
-        task: TaskType = TaskType.SEGMENTATION,
+        normalization: NORMALIZATION = NormalizationMethod.MIN_MAX,
+        threshold: THRESHOLD = "F1AdaptiveThreshold",
+        task: TaskType | str = TaskType.SEGMENTATION,
         image_metrics: str | list[str] | None = None,
         pixel_metrics: str | list[str] | None = None,
         visualizers: BaseVisualizer | list[BaseVisualizer] | None = None,
@@ -147,9 +139,14 @@ class Engine:
         self._cache = _TrainerArgumentsCache(callbacks=[*callbacks], **kwargs)
         self.normalization = normalization
         self.threshold = threshold
-        self.task = task
-        self.image_metric_names = image_metrics
-        self.pixel_metric_names = pixel_metrics
+        self.task = TaskType(task)
+        self.image_metric_names = image_metrics if image_metrics else ["AUROC", "F1Score"]
+
+        # pixel metrics are only used for segmentation tasks.
+        self.pixel_metric_names = None
+        if self.task == TaskType.SEGMENTATION:
+            self.pixel_metric_names = pixel_metrics if pixel_metrics is not None else ["AUROC", "F1Score"]
+
         self.visualizers = visualizers
 
         self.save_image = save_image