+++ src/anomalib/models/__init__.py
@@ -20,6 +20,7 @@ from .image import (
     Dfkde,
     Dfm,
     Draem,
+    Dsr,
     EfficientAd,
     Fastflow,
     Ganomaly,
@@ -62,7 +63,7 @@ __all__ = [
 logger = logging.getLogger(__name__)
 
 
-def _convert_pascal_to_snake_case(pascal_case: str) -> str:
+def convert_pascal_to_snake_case(pascal_case: str) -> str:
     """Convert PascalCase to snake_case.
 
     Args:
@@ -81,7 +82,7 @@ def _convert_pascal_to_snake_case(pascal_case: str) -> str:
     return re.sub(r"(?<!^)(?=[A-Z])", "_", pascal_case).lower()
 
 
-def _convert_snake_to_pascal_case(snake_case: str) -> str:
+def convert_snake_to_pascal_case(snake_case: str) -> str:
     """Convert snake_case to PascalCase.
 
     Args:
@@ -110,7 +111,7 @@ def get_available_models() -> set[str]:
         >>> get_available_models()
         ['ai_vad', 'cfa', 'cflow', 'csflow', 'dfkde', 'dfm', 'draem', 'efficient_ad', 'fastflow', ...]
     """
-    return {_convert_pascal_to_snake_case(cls.__name__) for cls in AnomalyModule.__subclasses__()}
+    return {convert_pascal_to_snake_case(cls.__name__) for cls in AnomalyModule.__subclasses__()}
 
 
 def _get_model_class_by_name(name: str) -> type[AnomalyModule]:
@@ -128,7 +129,7 @@ def _get_model_class_by_name(name: str) -> type[AnomalyModule]:
     logger.info("Loading the model.")
     model_class: type[AnomalyModule] | None = None
 
-    name = _convert_snake_to_pascal_case(name).lower()
+    name = convert_snake_to_pascal_case(name).lower()
     for model in AnomalyModule.__subclasses__():
         if name == model.__name__.lower():
             model_class = model