+++ src/anomalib/pipelines/__init__.py
@@ -1,4 +1,0 @@
-"""Anomalib Pipelines."""
-
-# Copyright (C) 2023-2024 Intel Corporation
-# SPDX-License-Identifier: Apache-2.0